backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
sentence-transformers
chromadb
pypdf
python-docx
sqlalchemy
python-multipart
pydantic
httpx
python-dotenv
aiohttp
livekit
livekit-api
livekit-agents[codecs]
livekit-plugins-google
